requirements.txt
# AI and Language Processing
google-generativeai  # For Gemini AI API
markdown  # For markdown processing
beautifulsoup4  # For HTML parsing (bs4)
langdetect

# Text-to-Speech
edge-tts  # Microsoft Edge TTS service
pygame  # For audio playback
sounddevice

# Core Dependencies
requests  # HTTP client
numpy
google-cloud-aiplatform

